Fixed server ports collection, added improvements.
