+++ ansible/roles/test/files/saitests/pfc_asym.py
@@ -32,7 +32,7 @@ class PfcAsymBaseTest(ThriftInterfaceDataPlane):
     PACKET_ECN = 1
     PACKET_TTL = 64
     PACKET_LEN = 72
-    PACKET_NUM = 10000
+    PACKET_NUM = 100000
 
     def __init__(self):
         ThriftInterfaceDataPlane.__init__(self)
@@ -146,6 +146,7 @@ class PfcAsymOffOnTxTest(PfcAsymBaseTest):
         # 2. Verify that PFC frames are generated for lossless priorities
         for sp in self.server_ports:
             port_counters, queue_counters = sai_thrift_read_port_counters(self.client, port_list[int(sp['index'])])
+            self.log("\nLossless: port_counters = {}\n queue_counters = {}".format(port_counters, queue_counters))
             assert(port_counters[self.INGRESS_DROP] > 0)
             for p in self.lossless_priorities:
                 assert(port_counters[p + 2] > 0)
@@ -156,6 +157,7 @@ class PfcAsymOffOnTxTest(PfcAsymBaseTest):
         # Verify that PFC frames are not generated for lossy priorities
         for sp in self.server_ports:
             port_counters, queue_counters = sai_thrift_read_port_counters(self.client, port_list[int(sp['index'])])
+            self.log("\Lossy: port_counters = {}\n queue_counters = {}".format(port_counters, queue_counters))
             for p in self.lossy_priorities:
                 assert(port_counters[p + 2] == 0)
 
@@ -205,9 +207,10 @@ class PfcAsymOffRxTest(PfcAsymBaseTest):
         assert(port_counters[self.INGRESS_DROP] == 0)
 
         # 1. Verify that packets are not dropped on dst ports
-        # 2. Verify that packets are transmitted from from dst ports
+        # 2. Verify that packets are transmitted from dst ports
         for sp in self.server_ports:
             port_counters, queue_counters = sai_thrift_read_port_counters(self.client, port_list[int(sp['index'])])
+            self.log("\Lossy - EGRESS_DROP: port_counters = {}\n queue_counters = {}".format(port_counters, queue_counters))
             assert(port_counters[self.EGRESS_DROP] == 0)
             for p in self.lossy_priorities:
                 assert(port_counters[self.TRANSMITTED_PKTS] > 0)
@@ -217,11 +220,13 @@ class PfcAsymOffRxTest(PfcAsymBaseTest):
 
         # Verify that some packets are dropped on src port, which means that Rx queue is full
         port_counters, queue_counters = sai_thrift_read_port_counters(self.client, port_list[int(self.non_server_port['index'])])
+        self.log("\nLossless - INGRESS_DROP: port_counters = {}\n queue_counters = {}".format(port_counters, queue_counters))
         assert(port_counters[self.INGRESS_DROP] > 0)
 
         # Verify that some packets are dropped on dst ports, which means that Tx buffer is full
         for sp in self.server_ports:
             port_counters, queue_counters = sai_thrift_read_port_counters(self.client, port_list[int(sp['index'])])
+            self.log("\nLossless - EGRESS_DROP: port_counters = {}\n queue_counters = {}".format(port_counters, queue_counters))
             assert(port_counters[self.EGRESS_DROP] > 0)
