+++ tests/conftest.py
@@ -74,6 +74,12 @@ def pytest_addoption(parser):
     parser.addoption("--vrf_capacity", action="store", default=None, type=int, help="vrf capacity of dut (4-1000)")
     parser.addoption("--vrf_test_count", action="store", default=None, type=int, help="number of vrf to be tested (1-997)")
 
+    ############################
+    # pfc_asym options         #
+    ############################
+    parser.addoption("--server_ports_num", action="store", default=20, type=int, help="Number of server ports to use")
+
+
 @pytest.fixture(scope="session")
 def testbed(request):
     """
